Expand metadata
Adds various pieces of useful information to the apps' metainfo file,
for display in app store frontends such as the Flathub website and GNOME
Software. This includes various URLs, supported screen sizes, and
supported input controls.

Also adds an issue tracker link to the about window and a Flathub button
to the readme.